feat: [M3-7445] - Public IP Addresses Tooltip and LISH Display
