+++ packages/manager/src/features/Linodes/AccessTable.test.tsx
@@ -2,7 +2,7 @@ import { fireEvent } from '@testing-library/react';
 import * as React from 'react';
 
 import { linodeFactory } from 'src/factories';
-import { PUBLIC_IPS_UNASSIGNED_TOOLTIP_TEXT } from 'src/features/Linodes/PublicIpsUnassignedTooltip';
+import { PUBLIC_IP_ADDRESSES_TOOLTIP_TEXT } from 'src/features/Linodes/PublicIPAddressesTooltip';
 import { renderWithTheme } from 'src/utilities/testHelpers';
 
 import { AccessTable } from './AccessTable';
@@ -28,7 +28,7 @@ describe('AccessTable', () => {
 
     const publicIpsUnassignedTooltip = await findByRole('tooltip');
     expect(publicIpsUnassignedTooltip).toContainHTML(
-      PUBLIC_IPS_UNASSIGNED_TOOLTIP_TEXT
+      PUBLIC_IP_ADDRESSES_TOOLTIP_TEXT
     );
 
     copyButtons.forEach((copyButton) => {
